requirements.txt
matplotlib>=3.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
requests>=2.26.0
pillow>=9.0.0
GPUtil>=1.4.0
psutil>=5.9.0